aider: Add the `cloudevents` package with a version of at least `1.2.0` to the list of dependencies in `pyproject.toml`.
# Aider chat conversation:

USER: I added these *read-write* files: pyproject.toml

ASSISTANT: To use the CloudEvents SDK in the `userup` project, we need to add the `cloudevents` package to the project's dependencies in `pyproject.toml`. Here's the diff for the changes:

```diff
--- pyproject.toml
+++ pyproject.toml
@@ ... @@
 requires-python = ">=3.8"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]

-dependencies = []
+dependencies = [
+    "cloudevents>=1.2.0"
+]

 [project.urls]
 Homepage = "https://github.com/hillside-labs/userservice-python"
 Issues = "https://github.com/hillside-labs/userservice-python/issues"
```

This change adds the `cloudevents` package with a version of at least `1.2.0` to the list of dependencies. Make sure to run the appropriate package management command to update your environment after applying this change, such as `pip install -U cloudevents` or `poetry update` if you are using Poetry.